+++ app/composables/fetchers/lists.ts
@@ -59,7 +59,7 @@ export async function deleteUserList(list: Partial<UserList> & { id: number }) {
 }
 
 export function getUserList(cid: number): UserList | null {
-    return useStore().user?.lists.find(x => x.users.some(x => x.cid === cid)) ?? null;
+    return useStore().user?.lists.find(x => x.users.some(x => x.cid === cid && !x.private)) ?? null;
 }
 
 export function sortList(users: UserListLiveUser[]) {