+++ prisma/migrations/20250720150539_user_private_mode/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE `User` ADD COLUMN `privateMode` BOOLEAN NOT NULL DEFAULT false,
+    ADD COLUMN `privateUntil` DATETIME(3) NULL;